# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/medicines/": {
            "get": {
                "tags": [
                    "Medicines"
                ],
                "summary": "Get list of medicines",
                "description": "Returns a list of medicines",
                "operationId": "listMedicines",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Medicines"
                ],
                "summary": "medicines  create",
                "description": "medicines  Customer",
                "operationId": "medicines/store",
                "requestBody": {
                    "description": "medicines create",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "generic_name",
                                    "buying_price",
                                    "stock_type",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "description": " name",
                                        "type": "string"
                                    },
                                    "generic_name": {
                                        "description": " generic_name",
                                        "type": "string"
                                    },
                                    "buying_price": {
                                        "description": " buying_price",
                                        "type": "interger"
                                    },
                                    "selling_price": {
                                        "description": " selling_price",
                                        "type": "interger"
                                    },
                                    "stock_type": {
                                        "description": " stock_type",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "stock": {
                                        "description": " stock",
                                        "type": "interger"
                                    },
                                    "category_id": {
                                        "description": " category_id",
                                        "type": "interger"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/medicines/{id}": {
            "get": {
                "tags": [
                    "Medicines"
                ],
                "summary": " medicines",
                "description": "medicines",
                "operationId": "medicines/get{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pass id of medicines",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Medicines"
                ],
                "summary": "medicines  update",
                "description": "medicines  update",
                "operationId": "medicines/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pass id for category",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "medicines create",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "generic_name",
                                    "buying_price",
                                    "stock_type",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "description": " name",
                                        "type": "string"
                                    },
                                    "generic_name": {
                                        "description": " generic_name",
                                        "type": "string"
                                    },
                                    "buying_price": {
                                        "description": " buying_price",
                                        "type": "interger"
                                    },
                                    "selling_price": {
                                        "description": " selling_price",
                                        "type": "interger"
                                    },
                                    "stock_type": {
                                        "description": " stock_type",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "stock": {
                                        "description": " stock",
                                        "type": "interger"
                                    },
                                    "category_id": {
                                        "description": " category_id",
                                        "type": "interger"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Medicines"
                ],
                "summary": "medicines detete",
                "description": "delete  medicines",
                "operationId": "medicines/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pass id for medicines",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Failed operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}